% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_gradient}
\alias{check_gradient}
\title{Gradient checking}
\usage{
check_gradient(network, epsilon = 1e-07)
}
\arguments{
\item{network}{\code{Network}: The network to be checked.}

\item{epsilon}{\code{numeric}: The (small) value used for calculating the approximated derivatives by stressing each parameter
up and down.}
}
\value{
\code{numeric}: The normalized L2-distance between the calculated/approximated derivatives.
}
\description{
Checks if the calculated derivatives are approximately equal to the approximated derivatives. If difference
is close to zero (smaller than \code{1e-6}), the backward pass is likely to be implemented correctly. Larger values may indicate
errors in the implementation.
}
