% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layer.R
\name{Layer}
\alias{Layer}
\title{Layer of a neural net}
\description{
The Layer class contains all functionalities needed (e.g. calculate the layer, calculate the derivative
of its weight matrix, etc.) for its use in the \link{Network} class.
One layer is not able to function on its own (aside from the input layer), the class is intentionally not exported.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Layer$new()}}
\item \href{#method-calc_layer}{\code{Layer$calc_layer()}}
\item \href{#method-calc_derivative}{\code{Layer$calc_derivative()}}
\item \href{#method-get_number_of_nodes}{\code{Layer$get_number_of_nodes()}}
\item \href{#method-get_weights}{\code{Layer$get_weights()}}
\item \href{#method-set_weights}{\code{Layer$set_weights()}}
\item \href{#method-set_input}{\code{Layer$set_input()}}
\item \href{#method-clone}{\code{Layer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new layer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$new(
  number_of_nodes,
  activation_fun,
  prior_layer = NULL,
  input_data = NULL,
  random_init = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{number_of_nodes}}{\code{integer} that specifies the number of nodes in the layer.}

\item{\code{activation_fun}}{\code{character} that contains the activation function to be used. See details.}

\item{\code{prior_layer}}{\code{Layer} in case of a hidden or output layer, the previous layer.}

\item{\code{input_data}}{\verb{matrix<numeric>} in case of the layer being the input layer, a\code{matrix} with numeric entries that contains the input_data.}

\item{\code{random_init}}{\code{integer}: Initializes all weights and biases with this value. If \code{NULL} assigns random values between -0.1 and 0.1.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Denote \code{a}the output of the activation function and \code{z} its input. Currently, the following activation functions are supported:
\itemize{
\item \code{relu}: \eqn{a = max(0,z)}
\item \code{identity}: \eqn{a=z}
\item \code{sigmoid}: \eqn{a = \frac{1}{1+\exp(-z)}}
\item \code{gelu}: \eqn{z \cdot \Phi(z)}
\item \code{silu}: \eqn{a = \frac{z}{1+\exp(-z)}}
\item \code{softplus}: \eqn{ a = \log(1+\exp{z})}
}
}

\subsection{Returns}{
\code{Layer}: The class instance. Invisibly, for chaining.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calc_layer"></a>}}
\if{latex}{\out{\hypertarget{method-calc_layer}{}}}
\subsection{Method \code{calc_layer()}}{
Calculate the output of the current layer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$calc_layer()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\verb{matrix<numeric>} Either the input- (for an input layer) or the output of the activation function.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calc_derivative"></a>}}
\if{latex}{\out{\hypertarget{method-calc_derivative}{}}}
\subsection{Method \code{calc_derivative()}}{
Calculate the derivative of the current layer wrt to its weights.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$calc_derivative(w_times_delta = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{w_times_delta}}{\verb{matrix<numeric>} either an output of the \code{calc_derivative()} method of the next layer, or the
gradient of the cost function (if the current layer is the output layer).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\verb{list<matrix<numeric>>} A list with two elements:
\itemize{
\item \code{derivative}: The derivative for the weighting matrix.
\item \code{w_times_delta}: A matrix product needed for the calculation of the previous layer.
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_number_of_nodes"></a>}}
\if{latex}{\out{\hypertarget{method-get_number_of_nodes}{}}}
\subsection{Method \code{get_number_of_nodes()}}{
Get the number of nodes of the layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$get_number_of_nodes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{integer} The number of nodes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_weights"></a>}}
\if{latex}{\out{\hypertarget{method-get_weights}{}}}
\subsection{Method \code{get_weights()}}{
Get the weight matrix of the layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$get_weights()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\verb{matrix<numeric>} The weights of the current layer.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_weights"></a>}}
\if{latex}{\out{\hypertarget{method-set_weights}{}}}
\subsection{Method \code{set_weights()}}{
Set/overwrite the weight matrix of the layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$set_weights(weight_mat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{weight_mat}}{\code{matrix} with numeric entries.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_input"></a>}}
\if{latex}{\out{\hypertarget{method-set_input}{}}}
\subsection{Method \code{set_input()}}{
Set/overwrite the input data of the layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$set_input(input_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_data}}{\code{matrix} with numeric entries.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
